+++ composer.json
@@ -27,7 +27,8 @@
         "symfony/dom-crawler": "^7.0",
         "illuminate/support": "^10.0",
         "guzzlehttp/guzzle": "^7.8",
-        "geniusts/prayer-times": "^1.0"
+        "geniusts/prayer-times": "^1.0",
+        "islamic-network/prayer-times": "^1.0"
     },
     "require-dev": {
         "laravel/pint": "^1.14",
