+++ config/prayertime.php
@@ -3,29 +3,33 @@
 return [
     /**
      * Source of prayertime.
-     * options: 'kemenag', 'myquran.com', 'manual calculation', 'custom'
+     *
+     * Supported: 'kemenag', 'myquran.com', 'manual calculation', 'custom'
+     * default: 'manual calculation'
      */
     'source' => ENV('PRAYERTIME_SOURCE', 'manual calculation'),
 
     /**
+     * manual calculation source
+     *
+     * Supported: "praytimes.org, geniusts/prayer-times, islamic-network/prayer-times"
+     * default: 'geniusts/prayer-times'
+     */
+    'manual_source' => ENV('PRAYTIME_MANUAL_SOURCE', 'geniusts/prayer-times'),
+
+    /**
      * Base URI of prayertime
+     *
      * kemenag: https://bimasislam.kemenag.go.id/
      * myquran.com: https://api.myquran.com/v2/
-     * manual calculation:
-     * custom:
+     * manual calculation: `leave this empty`
+     * custom: `free to customize`
      */
     'base_uri' => ENV('PRAYTIME_BASE_URI'),
 
     /**
-     * when using source 'manual calculation' you can choose to use
-     * package geniusts/prayer-times to do the calculation or using this class
-     *
-     * @see \Jhonoryza\LaravelPrayertime\Support\Concerns\CalculationPrayerTime
-     */
-    'use_package_geniusts_prayer_times' => false,
-
-    /**
      * define your custom prayer time class if source is using custom
+     *
      * you can check this class as reference
      *
      * @see \Jhonoryza\LaravelPrayertime\Support\KemenagPrayerTime
