+++ src/Support/Concerns/Interface/PrayerTime.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Jhonoryza\LaravelPrayertime\Support\Concerns;
+namespace Jhonoryza\LaravelPrayertime\Support\Concerns\Interface;
 
 /**
  * @method getFromLongLatOnSpecificYear(float $latitude, float $longitude, int $year)
